+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/Krypton Toast/Controls Visuals/User Input/LTR/ProgressBar/VisualToastComboBoxUserInputWithProgressBarForm.cs
@@ -83,9 +83,8 @@ internal partial class VisualToastComboBoxUserInputWithProgressBarForm : VisualT
 
     private void UpdateLocation()
     {
-        //Once loaded, position the form, or position it to the bottom left of the screen with added padding
-        Location = _data.NotificationLocation ?? new Point(Screen.PrimaryScreen!.WorkingArea.Width - Width - 5,
-            Screen.PrimaryScreen.WorkingArea.Height - Height - 5);
+        // Once loaded, position the form, or default to bottom-right with DPI-scaled edge padding.
+        Location = _data.NotificationLocation ?? GetDefaultBottomRightLocation();
     }
 
     private void UpdateIcon()
@@ -99,14 +98,8 @@ internal partial class VisualToastComboBoxUserInputWithProgressBarForm : VisualT
         SetIcon(bitmap);
     }
 
-    private void ShowCloseButton()
-    {
-        CloseBox = _data.ShowCloseBox ?? false;
-
-        FormBorderStyle = CloseBox ? FormBorderStyle.Fixed3D : FormBorderStyle.FixedSingle;
-
-        ControlBox = _data.ShowCloseBox ?? false;
-    }
+    private void ShowCloseButton() =>
+        ApplyCloseBoxChrome(_data.ShowCloseBox ?? false);
 
     private void VisualToastNotificationComboBoxUserInputWithProgressBarForm_Resize(object? sender, EventArgs e)
     {
@@ -120,9 +113,11 @@ internal partial class VisualToastComboBoxUserInputWithProgressBarForm : VisualT
     {
         UpdateIcon();
 
-        UpdateLocation();
-
         ShowCloseButton();
+
+        ApplyBorderlessHeightPadding();
+
+        UpdateLocation();
 
         kbtnDismiss.Text = KryptonManager.Strings.ToastNotificationStrings.Dismiss;