+++ Source/Krypton Components/Krypton.Toolkit.Utilities/Components/Krypton Toast/Controls Visuals/User Input/RTL/ProgressBar/VisualToastDomainUpDownInputWithProgressBarRtlAwareForm.cs
@@ -82,9 +82,8 @@ internal partial class VisualToastDomainUpDownInputWithProgressBarRtlAwareForm :
 
     private void UpdateLocation()
     {
-        //Once loaded, position the form, or position it to the bottom left of the screen with added padding
-        Location = _data.NotificationLocation ?? new Point(Screen.PrimaryScreen!.WorkingArea.Width - Width - 5,
-            Screen.PrimaryScreen.WorkingArea.Height - Height - 5);
+        // Once loaded, position the form, or default to bottom-right with DPI-scaled edge padding.
+        Location = _data.NotificationLocation ?? GetDefaultBottomRightLocation();
     }
 
     private void UpdateIcon()
@@ -98,14 +97,8 @@ internal partial class VisualToastDomainUpDownInputWithProgressBarRtlAwareForm :
         SetIcon(bitmap);
     }
 
-    private void ShowCloseButton()
-    {
-        CloseBox = _data.ShowCloseBox ?? false;
-
-        FormBorderStyle = CloseBox ? FormBorderStyle.Fixed3D : FormBorderStyle.FixedSingle;
-
-        ControlBox = _data.ShowCloseBox ?? false;
-    }
+    private void ShowCloseButton() =>
+        ApplyCloseBoxChrome(_data.ShowCloseBox ?? false);
 
     private void itbDismiss_Click(object sender, EventArgs e) => Close();
 
@@ -121,9 +114,11 @@ internal partial class VisualToastDomainUpDownInputWithProgressBarRtlAwareForm :
     {
         UpdateIcon();
 
-        UpdateLocation();
-
         ShowCloseButton();
+
+        ApplyBorderlessHeightPadding();
+
+        UpdateLocation();
 
         kbtnDismiss.Text = KryptonManager.Strings.ToastNotificationStrings.Dismiss;